Back-end/requirements.txt
  
asgiref==3.4.1
botocore==1.20.111
Django==3.2.5
djangorestframework==3.12.4
djongo==1.3.6
et-xmlfile==1.1.0
future==0.18.2
jmespath==0.10.0
jsonfield==3.1.0
openpyxl==3.0.7
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.6
